+++ src/components/Search/index.css
@@ -1,0 +1,38 @@
+.search-form{
+    width: 100%;
+    display:flex;
+    flex-direction: row;
+    align-items: center;
+    background-color: rgba(240, 241, 245);
+    padding:5px;
+    gap:10px;
+    border-radius: 5px;
+    padding:10px;
+}
+
+.search-form img{
+    opacity: 0.5;
+    width: 15px;
+}
+   
+.search-form input{
+    width: 100%;
+    border:0;
+    background-color: transparent;
+}
+
+.search-form input:focus{
+    outline:0;
+}
+
+/*Mobile*/
+@media (width < 830px){
+    .search-form{
+        gap:0;
+        width: 100%;
+    }
+    .search-form input{
+        width:80%;
+    }
+
+}
