+++ src/components/Search/index.jsx
@@ -1,0 +1,35 @@
+import './index.css'
+import {useState, useEffect, useContext} from 'react'
+import QueryFiltersContext from '../../context/filtersContext';
+import SearchIcon from '../../assets/search-icon.svg'
+
+function Search() {
+    const {setFilter, getActiveFilter} = useContext(QueryFiltersContext)
+    const [mounted, setMounted] = useState(false)
+    const [searchingValue, setSearchingValue] = useState(getActiveFilter("search"))
+
+    //search the product when the user ends writting on the search form
+    useEffect(() => {
+        if(mounted){
+            let timeOut = setTimeout(() => setFilter({name: "search", value:searchingValue}), 500)
+            return () => clearTimeout(timeOut)
+        }
+        else{
+            setMounted(true)
+        }
+    },[searchingValue])
+
+    return (
+        <form onSubmit={(e) => e.preventDefault()} className = "search-form">
+        <img src={SearchIcon} />
+        <input
+            placeholder="Buscar"
+            onChange={(e) => setSearchingValue(e.target.value)}
+            value={searchingValue}
+            style={{ color: "black" }}
+        />
+        </form>
+    );
+}
+
+export default Search;
