+++ src/components/SearchProduct/index.jsx
@@ -1,8 +1,8 @@
 import './index.css'
-import SearchIcon from '../../assets/search-icon.svg'
 import { Dropdown } from 'primereact/dropdown';
 import {useContext, useEffect, useState} from 'react'
 import QueryFiltersContext from '../../context/filtersContext';
+import Search from '../Search';
 
 const orderingValues = [
     {code: "", name : "Ordenar"},
@@ -17,38 +17,17 @@ const orderingValues = [
 function SearchProduct() {
     const {setFilter, getActiveFilter} = useContext(QueryFiltersContext)
     const [ordering, setOrdering] = useState(orderingValues.find(value => value.code === getActiveFilter("ordering")))
-    const [mounted, setMounted] = useState(false)
-    const [searchingValue, setSearchingValue] = useState(getActiveFilter("search"))
 
     function handleSetOrdering(value){
         setFilter({name: "ordering", value:value.code})
         setOrdering(value)
     }
 
-    //search the product when the user ends writting on the search form
-    useEffect(() => {
-        if(mounted){
-            let timeOut = setTimeout(() => setFilter({name: "search", value:searchingValue}), 500)
-            return () => clearTimeout(timeOut)
-        }
-        else{
-            setMounted(true)
-        }
-    },[searchingValue])
-
     return ( 
         <section className = "search-product">
             <h3>Productos</h3>
             <section className = "search-order-container">
-                <form onSubmit={(e) => e.preventDefault()}>
-                    <img src = {SearchIcon}/>
-                    <input 
-                        placeholder='Buscar' 
-                        onChange={(e) => setSearchingValue(e.target.value)} 
-                        value = {searchingValue}
-                        style={{color:"black"}}
-                        />
-                </form>
+                <Search/>
                 <Dropdown 
                     value={ordering} 
                     onChange={(e) => handleSetOrdering(e.value)}
